Refactored FittingWidget class from `fitting_widget_refactor` branch, updated for ver. 6.0 release source and with PR raised issues addressed.
